requirements.txt
# Chess Engine e Motor
python-chess==1.999
stockfish==3.28.0

# Chess.com API Integration
chessdotcom==2.1.1
requests==2.31.0

# Machine Learning
tensorflow==2.13.0
torch==2.1.0
scikit-learn==1.3.0
stable-baselines3==2.0.0
gym==0.29.1

# Data Handling
pandas==2.0.3
numpy==1.24.3
sqlite3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.13.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Optional Web Automation (se necessário)
selenium==4.11.2
playwright==1.37.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.0

